refactor: extracts shared types, schemas, and utilities to src/shared/
